requirements.txt
# Bu dosya, projenin gereksinimlerini içerir.
# Herhangi bir gereksinim eklemek için bu dosya güncellenebilir.
# Kullanım kolaylığı için vardır. Kod içinde aktif olarak çalışmaz.
# Gereksinimleri yüklemek için terminalde şu komutu çalıştırın:
# pip install -r requirements.txt

# Minimal gereksinimler
langchain
langchain-huggingface
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
llama-cpp-python
transformers
torch
python-dotenv

# Tüm ortam için → full_requirements.txt